==> 03_find_spicy_segments.ipynb <==
import pandas as pd
import re
import utils
import ast
import swifter
import json
from spicy_words import *

df_eu = pd.read_csv("xandr_segments_eu.csv") # load data
df_eu["name_list"] = df_eu["name_list"].apply(ast.literal_eval)


names = [
    'financial',
    'personality',
    'health',
    'sexuality',
    'race',
    'religion',
    'political',
    'children'
]

for name in names:
    word_list = eval(name)   
   
    spicy_re = utils.list_to_regex(word_list)

    spicy_dict = {word: [] for word in word_list}

    for index, row in df_eu.iterrows():
        for word in word_list:
            segname = row["name"]
            provider = row["provider_name"]
            if res := re.search(word.replace(" ", utils.word_separator), segname):
                spicy_dict[word].append(segname)

    filename = f"spicy_words_segmented/{name}.json"
    with open(filename, "w") as f:
        json.dump(spicy_dict, f, indent=4)
    print(filename)
